LPS-portal-service Inline ModuleResourceLoader. Only ServiceConfigurationInitializer invokes it.
